[svn-r12928] code revision

+++ tools/lib/h5trav_table.c
@@ -32,7 +32,7 @@
 
 int trav_table_search(haddr_t objno, trav_table_t *table )
 {
- int i;
+ unsigned int i;
 
  for (i = 0; i < table->nobjs; i++)
   if (table->objs[i].objno == objno)
@@ -61,7 +61,7 @@ void trav_table_add(haddr_t objno,
                     H5G_obj_t type,
                     trav_table_t *table)
 {
- int i;
+ unsigned int i;
 
  if (table->nobjs == table->size) {
   table->size *= 2;
@@ -110,7 +110,7 @@ void trav_table_addflags(unsigned *flags,
                          H5G_obj_t type,
                          trav_table_t *table)
 {
- int i;
+ unsigned int i;
 
  if (table->nobjs == table->size) {
   table->size *= 2;
@@ -157,7 +157,7 @@ void trav_table_addflags(unsigned *flags,
 
 void trav_table_init( trav_table_t **tbl )
 {
- int i;
+ unsigned int i;
  trav_table_t* table = (trav_table_t*) HDmalloc(sizeof(trav_table_t));
 
  table->size = 20;
@@ -197,7 +197,7 @@ void trav_table_init( trav_table_t **tbl )
 
 void trav_table_free( trav_table_t *table )
 {
- int i, j;
+ unsigned int i, j;
 
  for ( i = 0; i < table->nobjs; i++)
  {
@@ -234,14 +234,14 @@ void trav_table_addlink(trav_table_t *table,
                         int j /* the object index */,
                         char *path )
 {
- int k;
+ unsigned int k;
 
  /* already inserted */
  if (strcmp(table->objs[j].name,path)==0)
   return;
 
  /* allocate space if necessary */
- if (table->objs[j].nlinks == table->objs[j].sizelinks) {
+ if (table->objs[j].nlinks == (unsigned)table->objs[j].sizelinks) {
   table->objs[j].sizelinks += 2;
   table->objs[j].links =
    (trav_link_t*)HDrealloc(table->objs[j].links,
